Add initial work for assembler/disassembler
The pros of this approach is that its fairly simple to do.
The downside of this approach is that it doesnt really encapsulate
the operations. Ideally there would be a domain object for each operation
and then methods to map to or from a assembly nemonic or binary opcode

I'm going to refactor it to a more class-based approach later