Added code for assembling and disassembling chip8 instructions
Readme still needs to be updated, but its pretty much done.

The code in the Lib folder is a bit silly (its a 1500 line file), but
it contains 2 classes for each op code. I guess it could be split out,
but I don't know if its really that important.

Test cases are mostly working though I broke some stuff to support some of
the roms, which put random garbage at the end of the file (which I need to decode then recode).

I ran the disassembler against all the roms I have and they're mostly consistent, though there's
occasionally (spec non-complaint?) weird SRE op-codes which are having the last nibble replaced
with a 0 after decode/encode.